requirements.txt
pandas==1.5.3
pandas_gbq==0.21.0
protobuf==4.25.2
streamlit==1.31.0
streamlit-lottie==0.0.5
google-cloud-bigquery==3.17.2
google-auth==2.27.0
geopandas==0.14.3